Hayden branch

+++ hatServer/sortingHat/buildDB.py
@@ -3,7 +3,6 @@ import sys
 sys.path.append("..")
 sys.path.append("../..")
 from hatServer.models import Groups, Students
-import hatServer.sortingHat.leadership as l
 
 class dataSet:
     """
@@ -20,24 +19,47 @@ def parseStudent(data):
         identikey=values[2],
         known_skills=values[3].split('.'),
         learn_skills=values[4].split('.'),
-        leadership=values[5]
+        leadership=values[5],
+        ip_pref=values[6]
     )
     student_to_add.save()
     prefs = []
-    for i in range(6, 11):
+    for i in range(7, 12):
+#    for i in range(6, 9):
         g_name = values[i].rstrip('\n')
-        #group_to_add = groups.objects(group_name=g_name)
-        student_to_add.update(add_to_set__preferences=g_name)
-        #print(group_to_add)
+        group_to_add = Groups.objects.get(group_name=g_name)
+        student_to_add.update(add_to_set__preferences=group_to_add)
+        #print(group_to_add.group_name)
     student_to_add.save()
     return student_to_add
+
+def addWorkWith(data):
+    values = data.split(',')
+    student = Students.objects.get(identikey=values[2])
+    ids = values[12].split('.')
+    for identikey in ids:
+        if identikey == '0':
+            break
+        student_to_work_with = Students.objects.get(identikey=identikey)
+        student.update(add_to_set__work_with=student_to_work_with)
+    bad_ids = values[13].rstrip('\n')
+    bad_ids = bad_ids.split('.')
+    for identikey in bad_ids:
+        if identikey == '0':
+            break
+        student_to_avoid = Students.objects.get(identikey=identikey)
+        student.update(add_to_set__dont_work_with=student_to_avoid)
+    student.save()
 
 def parseGroups(data):
     values = data.split(',')
     group_to_add = Groups(group_name=values[0])
     group_to_add.save()
+    paid = bool(int(values[1]))
+    group_to_add.update(paid=paid)
+    group_to_add.update(ip=values[2])
     skills = []
-    for i in range(2, len(values)):
+    for i in range(3, len(values)):
         group_to_add.update(add_to_set__skills=values[i])
 #    group_to_add.update(remove_from_set__skills='\n')
 #    skills.remove('\n')
@@ -50,6 +72,10 @@ def parseGroups(data):
 
 def buildDB(student_path, group_path):
     #assert(len(student_path) > 0)
+    if len(Groups.objects.all()) > 0:
+        Groups.drop_collection()
+    if len(Students.objects.all()) > 0:
+        Students.drop_collection()
     student_data = dataSet(student_path)
     group_data = dataSet(group_path)
     for line in group_data.readData:
@@ -57,11 +83,16 @@ def buildDB(student_path, group_path):
         # TODO add proper mongoengine command here if this doesn't work
     for line in student_data.readData:
         s = parseStudent(line)
+    for line in student_data.readData:
+        s_with = addWorkWith(line)
 
 def main(args):
-    db = connect('testDB')
-    Groups.drop_collection()
-    Students.drop_collection()
+    register_connection('testDatabase')
+    connect('testDatabase')
+    if len(Groups.objects.all()) > 0:
+        Groups.drop_collection()
+    if len(Students.objects.all()) > 0:
+        Students.drop_collection()
     buildDB(args[1], args[2])
 
 if __name__ == '__main__':
